1) poprawka video - problem konwersji float do double 2) dodanie wstęg do wykresu 2D 3) dodanie splash screen 4) dodanie zaślepki logowania i rejestracji (login: k.lebek@polsl.pl) 5) poprawki w plugin_subject - dodanie namespace
git-svn-id: https://localhost:8443/svn/hm-mde-1@643 fca3399c-2208-499e-aaa1-2cf7fb07c25e

+++ trunk/src/edr/include/core/IVisualizer.h
@@ -64,16 +64,16 @@ namespace core
 
             virtual ~TimeSerieBase() {}
 
-            virtual void setTime(float time) = 0;
+            virtual void setTime(double time) = 0;
 
-            virtual float getLength() const = 0;
+            virtual double getLength() const = 0;
 
-            virtual void setOffset(float offset)
+            virtual void setOffset(double offset)
             {
 
             }
 
-            virtual void setScale(float scale)
+            virtual void setScale(double scale)
             {
 
             }
